refactor(oauth): split oauth.rs into per-concern modules
oauth.rs had grown to ~2600 lines interleaving three concerns: the
generic OAuth trait/registry/device-flow machinery, the GitHub Copilot
provider, and the Anthropic provider. Split it into a module:

  oauth/mod.rs            - shared types, provider registry, generic
                            device-code polling, domain helpers, and
                            #[cfg(test)] test_support helpers
  oauth/github_copilot.rs - GitHub Copilot provider + tests
  oauth/anthropic.rs      - Anthropic provider (PKCE, callback server) + tests

Pure move: no behavior change. Public API is preserved via re-exports
(`pub use {anthropic,github_copilot}::*`), and the two pub(crate) helpers
used by the streaming provider module are re-exported explicitly. All
382 tests pass; clippy clean.
